fixed shared memory max value overflow. changed to  unsigned long

+++ postgresql-9.2.1/src/include/miscadmin.h
@@ -229,7 +229,7 @@ extern int	CTimeZone;
 extern bool enableFsync;
 extern bool allowSystemTableMods;
 extern PGDLLIMPORT int work_mem;
-extern PGDLLIMPORT unsigned long smooth_work_mem;
+extern PGDLLIMPORT int smooth_work_mem;
 extern PGDLLIMPORT int maintenance_work_mem;
 
 extern int	VacuumCostPageHit;
